v3/requirements.txt
anthropic>=0.15.0
rich>=13.0.0
python-dotenv>=1.0.0